creative-cli/requirements.txt
# Core dependencies
click==8.1.7
pyyaml==6.0.1
python-dotenv==1.0.0

# Image processing
Pillow==10.3.0
opencv-python==4.9.0.80

# OpenAI for GenAI image generation
openai==1.12.0

# HTTP requests
requests==2.31.0

# Logging and utilities
colorama==0.4.6
tqdm==4.66.1

# Validation
jsonschema==4.21.1

# Optional: For better image quality
numpy==1.26.4
